backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
yfinance==0.2.36
numpy==1.26.3
